Refactor k8s informer for cluster
